Change quantity from int to long

+++ app/src/main/java/itemsale/suvidha/com/itemsale/features/newsale/NewSaleViewModel.kt
@@ -23,7 +23,7 @@ class NewSaleViewModel(
   data class ViewState(
     val items: List<Item>,
     val isPaid: Boolean,
-    val totalQuantity: Int,
+    val totalQuantity: Long,
     val subTotal: Double,
     val isDone: Boolean,
     val isGreaterAmount: Boolean,
